+++ src/builder/Block.kt
@@ -1,13 +1,15 @@
 package builder
 
 class Block(val name: String, val texture: String, val hardness: Double, val resistance: Double, val unbreakable: Boolean,
-            val destroyedByExplosion: Boolean, val climbable: Boolean, val slipperiness: Double, val light: Double,
-            val tool: String, val harvestLevel: Int, val overrides: Map<String, String>) {
+            val destroyedByExplosion: Boolean, val flammable: Boolean, val burnForever: Boolean, val redstoneSignal: Int,
+            val slipperiness: Double, val light: Double, val tool: String, val harvestLevel: Int, val gravity: Boolean,
+            val overrides: Map<String, String>) {
 
     fun registryName(): String = name.replace(" ", "_").toLowerCase()
 
     fun toJava(): String {
-        var java = "VulcanMod.blocks.addBlock(\"${registryName()}\", new VulcanBlock($hardness, $resistance, $unbreakable, $slipperiness, $light, \"$tool\", $harvestLevel, $destroyedByExplosion, $climbable, null, 0, 0)"
+        var blockClass = if(gravity) "VulcanBlockFalling" else "VulcanBlock"
+        var java = "VulcanMod.blocks.addBlock(\"${registryName()}\", new $blockClass($hardness, $resistance, $unbreakable, $slipperiness, $light, \"$tool\", $harvestLevel, $destroyedByExplosion, $flammable, $burnForever, $redstoneSignal, null, 0, 0)"
         if(overrides.isNotEmpty()) {
             java += " {¶"
             overrides.keys.asSequence().forEach {